Style footer section
Add content-container div to wrap footer-text, in case the footer
content is later expanded.

To-Do:
Need to rename content-container to content-alignment, as the name
was mistakenly given. Separate containers for alignment need to be
used for the quote, information, and call to action so that the
hero section, call to action container, and header are wider than
that content.